refactor(web): dissolve lib/conversations-api.ts into domain modules (LUM-1982)
Decompose the 428-line lib/conversations-api.ts into 6 targeted modules
per CONVENTIONS.md directory placement rules:

- types/conversation-types.ts — Conversation, ConversationChannelBinding,
  ConversationSlackChannel, ConversationSlackThread, ConversationGroup
- types/subagent-types.ts — SubagentEvent, SubagentDetailResponse
- domains/conversations/conversation-transforms.ts — toConversation,
  detailToConversation, RawConversationSummary, RawConversationDetail
- domains/conversations/fetch-conversation-detail.ts —
  fetchConversationDetail, CONVERSATION_NOT_FOUND
- utils/conversation-predicates.ts — isBackgroundConversation,
  canMarkUnread, canMarkRead
- utils/fetch-subagent-detail.ts — fetchSubagentDetail wrapper

Inline fetchConversationList and listConversations into
conversation-queries.ts. Update 38 consumer import paths.
Delete lib/conversations-api.ts.

Closes LUM-1982
