+++ apps/web/src/domains/chat/components/assistant-side-menu.test.tsx
@@ -18,7 +18,7 @@ mock.module("@/hooks/use-is-mobile", () => ({
 }));
 
 
-import type { Conversation } from "@/lib/conversations-api";
+import type { Conversation } from "@/types/conversation-types";
 import {
   ASSISTANT_SIDE_MENU_CONVERSATION_LIMIT,
   AssistantSideMenu,
