+++ apps/web/src/domains/chat/components/slack-channel-footer.tsx
@@ -6,7 +6,7 @@ import { resolveSlackChannelName } from "@/domains/chat/api/slack-channel-name";
 import type {
   Conversation,
   ConversationChannelBinding,
-} from "@/lib/conversations-api";
+} from "@/types/conversation-types";
 import {
   getSlackLinkUrl,
   type DisplayMessage,
