+++ apps/web/src/domains/chat/components/sub-group-accordion.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from "react";
 
 import { CollapsibleNavSection } from "@/components/collapsible-nav-section";
 import { PanelItem, SideMenu } from "@vellum/design-library";
-import type { Conversation } from "@/lib/conversations-api";
+import type { Conversation } from "@/types/conversation-types";
 import type { SubGroup } from "@/domains/chat/utils/sub-group-utils";
 import { countBadge } from "@/domains/chat/components/sidebar-count-badge";
